[wasm][Driver] HACK: Add wasi-sysroot layout and auto-detection support
Currenty, SwiftWasm toolchain contains wasi-sysroot under
`./usr/share/wasi-sysroot` directory in the toolchain package, but we
don't have a good consensus on the location of wasi-sysroot yet. So this
patch is a temporary solution just to avoid the need to specify `-sdk`
option manually. That being said, SwiftPM already passes `-sdk` option
to the driver, so the option is only needed when the driver is invoked
directly.