[test] Fix calling convention mismatch for swift_demangle
`test/stdlib/Runtime.swift.gyb` calls `swift_demangle` with the Swift
calling convention, but the functions is defined with the C calling
convention. But this does not cause a problem other than testing, and
ideal fix would be to put `@_cdecl` on the function declared in
Swift-side to choose the correct calling convention. `@cdecl` without
body is not supported yet but discussed in https://forums.swift.org/t/formalizing-cdecl/40677