[stdlib] Fix calling convention mismatch for swift_COWChecksEnabled
`ContiguousArrayBuffer.swift` calls `swift_COWChecksEnabled` with the
Swift calling convention, but the functions is defined with the C calling
convention. This mismatch actually happens only when building with
`enable-array-cow-checks` (which is not enabled by default).
Also ideal fix would be to put `@_cdecl` on the function declared in
Swift-side to choose the correct calling convention. `@cdecl` without
body is not supported yet but discussed in https://forums.swift.org/t/formalizing-cdecl/40677