[stdlib] Fix calling convention mismatch for runtime function tracking
Runtime function tracking APIs are defined with the C calling convention
but they are called from Swift through `@_silgen_name`, so they are
called with the Swift calling convention. This patch fixes the mismatch
by adding `SWIFT_CC(swift)` to the definitions, but ideally we should be
able to call these APIs with the C calling convention with `@_cdecl`.
`@cdecl` without body is not supported yet but discussed in
https://forums.swift.org/t/formalizing-cdecl/40677