Set git credentials by git-config instead of depending on actions/checkout
Because swift repository is now checkout under ../build/Distribution by
gh-distribute-toolchain and the path can be changed in the future, we
shouldn't assume the path of swift repository. actions/checkout strategy
to pass credentials to git depends on the path, so we manually set the
credentials by ourselves.
The `extraheader` strategy somehow doesn't work now, so we use basic
authentication embedded in the url.