[wasm] Mark `errno` as nonisolated(unsafe) in wasi-libc
This patch is a workaround to reflect the fact that `errno` is thread-local
and can be accessed from any actor. This is a temporary solution until we
have `__swift_attr__` support in apinotes or `thread_local` support in
ClangImporter.

This change is required to build swift-corelibs-foundation with Swift 6
language mode.